Add update method to DbManager
Add catch for the create method. If a player with the email already exists in the database, we print this to stdout but the program doesn't crash.

Fix null pointer where we returned Optional.of(player) instead of Optional.empty() when reading a record that doesn't exist in the db.